requirements.txt
Django>=5.2.10,<6.0
dj-database-url==3.1.2
django-constance==4.3.4
psycopg2-binary==2.9.11
python-dotenv==1.2.2
whitenoise==6.12.0
gunicorn==25.3.0
requests==2.32.5
httpx==0.28.1
anthropic>=0.39.0
